fix: address PR review comment
Move shared LABEL fix helpers into labels/util.go and reuse the shared Buildx label key constant.

Addresses: https://github.com/wharflab/tally/pull/622#discussion_r3190998432
